+++ src/js/fake-translator.js
@@ -27,11 +27,14 @@ var SOURCE_STRINGS = {
         'placeholder': 'Click here to upload file. (< __maxSize__)',
         'notFound': 'File __existing__ could not be found (leave unchanged if already submitted and you want to preserve it).',
         'waitingForPermissions': 'Waiting for user permissions.',
-        'resetWarning': 'This will remove the file. Are you sure you want to do this?',
-        'toolargeerror': 'File too large (> __maxSize__)'
+        'resetWarning': 'This will remove the __item__. Are you sure you want to do this?',
+        'toolargeerror': 'File too large (> __maxSize__)',
+        'file': 'file'
     },
     'drawwidget': {
-        'resetWarning': 'This will remove the drawing, signature and/or file. Are you sure you want to do this?'
+        'drawing': 'drawing',
+        'signature': 'signature',
+        'annotation': 'file and drawing'
     },
     'form': {
         'required': 'required'